+++ gui/gui.h
@@ -35,6 +35,6 @@ void Gui_Destroy(void);
  * @param  data_len: data lenght in bytes
  * @retval None
  */
-void Gui_SendMessage(gui_message_type_t message_type, uint8_t * data, uint16_t data_len);
+void Gui_SendMessage(gui_message_type_t message_type, float * data, int data_len);
 
 #endif /* GUI_H */
